restructure test fixtures and directory layout
