--- tests/fixtures/__init__.py
@@ -1,10 +1,0 @@
-from .fixtures import webapi_client
-from .fixtures import mocked_renderer_device
-from .fixtures import mocked_server_device
-from .fixtures import mocked_av_control_point
-#from .fixtures import mediarenderer_http_responses
-from .fixtures import upnp_test_requester
-
-__all__ = [
-    'webapi_client', 'mocked_av_control_point', 'upnp_test_requester', 'mocked_renderer_device', 'mocked_server_device'
-]
