+++ tests/testsupport/advertisement_data.py
